master - fixed compiler warnings

+++ src/meta/DrawInterface.cpp
@@ -183,12 +183,12 @@ void ReyEngine::minimizeWindow() {
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////
-template<> Circle ReyEngine::Rect<double>::circumscribe() {return {{x + width / 2, y+height/2}, height/0.70710678118/2};}
-template<> Circle ReyEngine::Rect<int>::circumscribe() {return {{x + width / 2, y+height/2}, height/0.70710678118/2};}
-template<> Circle ReyEngine::Rect<float>::circumscribe() {return {{x + width / 2, y+height/2}, height/0.70710678118/2};}
-template<> Circle ReyEngine::Rect<double>::inscribe() {return {{x + width / 2, y+height/2}, (double)height/2};}
-template<> Circle ReyEngine::Rect<int>::inscribe() {return {{x + width / 2, y+height/2}, (double)height/2};}
-template<> Circle ReyEngine::Rect<float>::inscribe() {return {{x + width / 2, y+height/2}, (double)height/2};}
+template<> Circle ReyEngine::Rect<double>::circumscribe() {return {{(double)(x + width / 2), (double)(y+height/2)}, (double)height/0.70710678118/2};}
+template<> Circle ReyEngine::Rect<int>::circumscribe() {return {{(double)(x + width / 2), (double)(y+height/2)}, (double)height/0.70710678118/2};}
+template<> Circle ReyEngine::Rect<float>::circumscribe() {return {{(double)(x + width / 2), (double)(y+height/2)}, (double)height/0.70710678118/2};}
+template<> Circle ReyEngine::Rect<double>::inscribe() {return {{(double)(x + width / 2), (double)(y+height/2)}, (double)height/2};}
+template<> Circle ReyEngine::Rect<int>::inscribe() {return {{(double)(x + width / 2), (double)(y+height/2)}, (double)height/2};}
+template<> Circle ReyEngine::Rect<float>::inscribe() {return {{(double)(x + width / 2), (double)(y+height/2)}, (double)height/2};}
 
 /////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////
